requirements/base.txt
amqp==5.3.1
arabic-reshaper==3.0.0
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
binaryornot==0.4.4
blinker==1.9.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cookiecutter==2.6.0
cryptography==43.0.3
cssselect2==0.7.0
db-dtypes==1.3.1
Django==5.1.3
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
django-htmx==1.21.0
django-otp==1.5.4
django-redis==5.4.0
django-tailwind==3.8.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
flasgger==0.9.7.1
Flask==3.1.0
flask-sock==0.7.0
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaleido==0.2.1
kombu==5.4.2
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistune==3.0.2
msgpack==1.1.0
numpy==2.1.3
oscrypto==1.3.0
packaging==24.2
pandas==2.2.3
phonenumbers==8.13.50
pillow==11.0.0
plotly==5.24.1
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
PyJWT==2.10.0
pyOpenSSL==24.2.1
pypdf==5.1.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
redis==5.2.0
referencing==0.35.1
reportlab==4.2.5
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
service-identity==24.2.0
simple-websocket==1.1.0
six==1.16.0
SQLAlchemy==2.0.36
sqlparse==0.5.2
svglib==1.5.1
tabulate==0.9.0
tenacity==9.0.0
text-unidecode==1.3
tinycss2==1.4.0
Twisted==24.10.0
txaio==23.1.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.3
vanna==0.7.5
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
whitenoise==6.8.2
wsproto==1.2.0
xhtml2pdf==0.2.16
zope.interface==7.1.1
